Move some code to Java front-end
